+++ ta4j-core/src/main/java/org/ta4j/core/bars/AmountBarBuilder.java
@@ -36,43 +36,53 @@ import org.ta4j.core.num.Num;
 import org.ta4j.core.num.NumFactory;
 
 /**
- * A tick bar is sampled after a fixed number of ticks.
+ * An amount bar is sampled after a fixed number of amount (= price * volume)
+ * have been traded.
  */
-public class TickBarBuilder implements BarBuilder {
+public class AmountBarBuilder implements BarBuilder {
+
+    private Num distinctVolume;
+    private final Num amountThreshold;
+    private final boolean setAmountByVolume;
 
     private final NumFactory numFactory;
-    private final int tickCount;
-    private int passedTicksCount;
     private BarSeries barSeries;
     private Duration timePeriod;
     private Instant beginTime;
     private Instant endTime;
-    private Num volume;
     private Num openPrice;
     private Num highPrice;
-    private Num closePrice;
     private Num lowPrice;
+    private Num closePrice;
+    private Num volume;
     private Num amount;
     private long trades;
 
     /**
      * A builder to build a new {@link BaseBar} with {@link DoubleNumFactory}
      *
-     * @param tickCount the number of ticks at which a new bar should be created
+     * @param amountThreshold   the threshold at which a new bar should be created
+     * @param setAmountByVolume if {@code true} the {@link #amount} is set by
+     *                          {@link #volume} * {@link #closePrice}, otherwise
+     *                          {@link #amount} must be explicitly set
      */
-    public TickBarBuilder(final int tickCount) {
-        this(DoubleNumFactory.getInstance(), tickCount);
+    public AmountBarBuilder(final int amountThreshold, final boolean setAmountByVolume) {
+        this(DoubleNumFactory.getInstance(), amountThreshold, setAmountByVolume);
     }
 
     /**
      * A builder to build a new {@link BaseBar}
      *
      * @param numFactory
-     * @param tickCount  the number of ticks at which a new bar should be created
+     * @param amountThreshold   the threshold at which a new bar should be created
+     * @param setAmountByVolume if {@code true} the {@link #amount} is set by
+     *                          {@link #volume} * {@link #closePrice}, otherwise
+     *                          {@link #amount} must be explicitly set
      */
-    public TickBarBuilder(final NumFactory numFactory, final int tickCount) {
+    public AmountBarBuilder(final NumFactory numFactory, final int amountThreshold, final boolean setAmountByVolume) {
         this.numFactory = numFactory;
-        this.tickCount = tickCount;
+        this.amountThreshold = numFactory.numOf(amountThreshold);
+        this.setAmountByVolume = setAmountByVolume;
         reset();
     }
 
@@ -96,47 +106,47 @@ public class TickBarBuilder implements BarBuilder {
 
     @Override
     public BarBuilder openPrice(final Num openPrice) {
-        throw new IllegalArgumentException("TickBar can only be built from closePrice");
+        throw new IllegalArgumentException("AmountBar can only be built from closePrice");
     }
 
     @Override
     public BarBuilder openPrice(final Number openPrice) {
-        throw new IllegalArgumentException("TickBar can only be built from closePrice");
+        throw new IllegalArgumentException("AmountBar can only be built from closePrice");
     }
 
     @Override
     public BarBuilder openPrice(final String openPrice) {
-        throw new IllegalArgumentException("TickBar can only be built from closePrice");
+        throw new IllegalArgumentException("AmountBar can only be built from closePrice");
     }
 
     @Override
     public BarBuilder highPrice(final Number highPrice) {
-        throw new IllegalArgumentException("TickBar can only be built from closePrice");
+        throw new IllegalArgumentException("AmountBar can only be built from closePrice");
     }
 
     @Override
     public BarBuilder highPrice(final String highPrice) {
-        throw new IllegalArgumentException("TickBar can only be built from closePrice");
+        throw new IllegalArgumentException("AmountBar can only be built from closePrice");
     }
 
     @Override
     public BarBuilder highPrice(final Num highPrice) {
-        throw new IllegalArgumentException("TickBar can only be built from closePrice");
+        throw new IllegalArgumentException("AmountBar can only be built from closePrice");
     }
 
     @Override
     public BarBuilder lowPrice(final Num lowPrice) {
-        throw new IllegalArgumentException("TickBar can only be built from closePrice");
+        throw new IllegalArgumentException("AmountBar can only be built from closePrice");
     }
 
     @Override
     public BarBuilder lowPrice(final Number lowPrice) {
-        throw new IllegalArgumentException("TickBar can only be built from closePrice");
+        throw new IllegalArgumentException("AmountBar can only be built from closePrice");
     }
 
     @Override
     public BarBuilder lowPrice(final String lowPrice) {
-        throw new IllegalArgumentException("TickBar can only be built from closePrice");
+        throw new IllegalArgumentException("AmountBar can only be built from closePrice");
     }
 
     @Override
@@ -164,37 +174,41 @@ public class TickBarBuilder implements BarBuilder {
 
     @Override
     public BarBuilder volume(final Num volume) {
-        this.volume = this.volume.plus(volume);
+        this.distinctVolume = volume;
+        this.volume = this.volume == null ? volume : this.volume.plus(volume);
         return this;
     }
 
     @Override
     public BarBuilder volume(final Number volume) {
-        volume(this.numFactory.numOf(volume));
+        volume(numFactory.numOf(volume));
         return this;
     }
 
     @Override
     public BarBuilder volume(final String volume) {
-        volume(this.numFactory.numOf(volume));
+        volume(numFactory.numOf(volume));
         return this;
     }
 
     @Override
     public BarBuilder amount(final Num amount) {
+        if (setAmountByVolume) {
+            throw new IllegalArgumentException("AmountBar.amount can only be built from closePrice*volume");
+        }
         this.amount = this.amount == null ? amount : this.amount.plus(amount);
         return this;
     }
 
     @Override
     public BarBuilder amount(final Number amount) {
-        amount(this.numFactory.numOf(amount));
+        amount(numFactory.numOf(amount));
         return this;
     }
 
     @Override
     public BarBuilder amount(final String amount) {
-        amount(this.numFactory.numOf(amount));
+        amount(numFactory.numOf(amount));
         return this;
     }
 
@@ -211,7 +225,7 @@ public class TickBarBuilder implements BarBuilder {
     }
 
     @Override
-    public TickBarBuilder bindTo(final BarSeries barSeries) {
+    public AmountBarBuilder bindTo(final BarSeries barSeries) {
         this.barSeries = Objects.requireNonNull(barSeries);
         return this;
     }
@@ -229,25 +243,43 @@ public class TickBarBuilder implements BarBuilder {
 
     @Override
     public void add() {
-        if (++passedTicksCount % tickCount == 0) {
-            if (amount == null && volume != null) {
-                amount = closePrice.multipliedBy(volume);
+
+        if (setAmountByVolume) {
+            final var calculatedAmount = closePrice.multipliedBy(distinctVolume);
+            amount = amount == null ? calculatedAmount : amount.plus(calculatedAmount);
+        }
+
+        if (amount.isGreaterThanOrEqual(amountThreshold)) {
+            // move amount remainder to next bar
+            var amountRemainder = numFactory.zero();
+            // move volume remainder to next bar
+            var volumeRemainder = numFactory.zero();
+            if (amount.isGreaterThan(amountThreshold)) {
+                amountRemainder = amount.minus(amountThreshold);
+                volumeRemainder = amountRemainder.dividedBy(closePrice);
+
+                // cap currently built bar, amount is then restored to amountRemainder
+                amount = amountThreshold;
+                // cap currently built bar, volume is then restored to volumeRemainder
+                volume = volume.minus(volumeRemainder);
             }
 
             barSeries.addBar(build());
+            amount = amountRemainder;
+            volume = volumeRemainder;
+
             reset();
         }
     }
 
     private void reset() {
-        final var zero = numFactory.zero();
+        distinctVolume = null;
+
         timePeriod = null;
         openPrice = null;
-        highPrice = zero;
+        highPrice = numFactory.zero();
         lowPrice = numFactory.numOf(Integer.MAX_VALUE);
         closePrice = null;
-        amount = null;
         trades = 0;
-        volume = zero;
     }
 }