+++ ta4j-core/src/main/java/org/ta4j/core/indicators/MassIndexIndicator.java
@@ -24,11 +24,10 @@
 package org.ta4j.core.indicators;
 
 import org.ta4j.core.BarSeries;
-import org.ta4j.core.Indicator;
 import org.ta4j.core.indicators.averages.EMAIndicator;
-import org.ta4j.core.indicators.helpers.CombineIndicator;
 import org.ta4j.core.indicators.helpers.HighPriceIndicator;
 import org.ta4j.core.indicators.helpers.LowPriceIndicator;
+import org.ta4j.core.indicators.numeric.BinaryOperation;
 import org.ta4j.core.num.Num;
 
 /**
@@ -53,7 +52,7 @@ public class MassIndexIndicator extends CachedIndicator<Num> {
      */
     public MassIndexIndicator(BarSeries series, int emaBarCount, int barCount) {
         super(series);
-        Indicator<Num> highLowDifferential = CombineIndicator.minus(new HighPriceIndicator(series),
+        final var highLowDifferential = BinaryOperation.difference(new HighPriceIndicator(series),
                 new LowPriceIndicator(series));
         this.singleEma = new EMAIndicator(highLowDifferential, emaBarCount);
         this.doubleEma = new EMAIndicator(singleEma, emaBarCount); // Not the same formula as DoubleEMAIndicator