+++ CHANGELOG.md
@@ -4,6 +4,7 @@ Changelog for `ta4j`, roughly following [keepachangelog.com](http://keepachangel
 
 ### Breaking
 - Refactored `ProfitLossCriterion`, `ProfitCriterion`, `LossCriterion`, `AverageProfitCriterion`, `AverageLossCriterion`, `ReturnCriterion`, `ProfitLossRatioCriterion` and `ProfitLossPercentageCriterion` criteria into their net and gross concrete classes
+- [#1266](https://github.com/ta4j/ta4j/issues/1266) Consolidated BinaryOperation, UnaryOperation, TransformIndicator and CombineIndicator
 
 ### Fixed
 - Updated Github test workflow to cache dependencies for quicker builds
@@ -14,17 +15,20 @@ Changelog for `ta4j`, roughly following [keepachangelog.com](http://keepachangel
 - Fixed strict rules of `ConvergenceDivergenceIndicator`
 - Fixed calculation of `ReturnOverMaxDrawdownCriterion`
 - swapped parameter naming in  `BaseBarSeries#addTrade(final Number tradeVolume, final Number tradePrice)`
+- Aggregation of amount and trades in `VolumeBarBuilder` and `TickBarBuilder`
 
 ### Changed
 - Use `NetReturnCriterion` in `AverageReturnPerBarCriterion`, `EnterAndHoldCriterion` and `ReturnOverMaxDrawdownCriterion` to avoid optimistic bias of `GrossReturnCriterion`
 - `ReturnOverMaxDrawdownCriterion` now returns 0 instead of `NaN` for strategies that never operate, and returns the net profit instead of `NaN` for strategies with no drawdown
+- Changed snapshot distribution to Maven Central after OSSRH end-of-life
 
 ### Removed/Deprecated
+- TransformIndicator and CombineIndicator
 
 ### Added
 - Bars can now be built by `beginTime` instead of `endTime`
+- Added `AmountBarBuilder` to `bars`-package to aggregate bars after a fixed number of amount have been traded
 - Added `CumulativePnL` and `MaximumAbsoluteDrawdownCriterion` to calculate the max drawdown absolute value, and `MaxDrawdownBarLengthCriterion` to calculate its length
-- 
 
 ## 0.18 (released May 15, 2025)